Move Class Maps into their own folder
