+++ src/EPR.SubsidiaryBulkUpload.Application.UnitTests/Services/CsvProcessorTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Text;
+using EPR.SubsidiaryBulkUpload.Application.ClassMaps;
 using EPR.SubsidiaryBulkUpload.Application.CsvReaderConfiguration;
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Models;
