+++ src/EPR.SubsidiaryBulkUpload.Application/ClassMaps/CompaniesHouseCompanyMap.cs
@@ -1,9 +1,10 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using CsvHelper;
 using CsvHelper.Configuration;
+using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Models;
 
-namespace EPR.SubsidiaryBulkUpload.Application.DTOs;
+namespace EPR.SubsidiaryBulkUpload.Application.ClassMaps;
 
 [ExcludeFromCodeCoverage]
 public class CompaniesHouseCompanyMap : ClassMap<CompaniesHouseCompany>
