+++ src/EPR.SubsidiaryBulkUpload.Application/Services/ParserClass.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text;
 using CsvHelper.Configuration;
+using EPR.SubsidiaryBulkUpload.Application.ClassMaps;
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Models;
 using Microsoft.Extensions.Logging;
