+++ src/EPR.SubsidiaryBulkUpload.Function/BulkUploadFunction.cs
@@ -1,4 +1,5 @@
 ﻿using Azure.Storage.Blobs;
+using EPR.SubsidiaryBulkUpload.Application.ClassMaps;
 using EPR.SubsidiaryBulkUpload.Application.CsvReaderConfiguration;
 using EPR.SubsidiaryBulkUpload.Application.DTOs;
 using EPR.SubsidiaryBulkUpload.Application.Extensions;
